+++ fs-api/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for fs-api
 
-## next version
+## 0.2.0.0 -- 2023-08-01
 
 ### Breaking
 
@@ -8,7 +8,9 @@
   For this reason, we move the relevant modules into an `Internal` hierarchy.
   * Move the `System.IO.FS` module to `System.FS.IO.Internal`.
   * Move the `System.FS.Handle` module to `System.FS.IO.Internal.Handle`.
-* Move strict and lazy compound definitions for reading/writing bytes into separate modules `System.FS.API.Strict` and `System.FS.API.Lazy`. Both modules re-export `System.FS.API`.
+* Move strict and lazy compound definitions for reading/writing bytes into
+  separate modules `System.FS.API.Strict` and `System.FS.API.Lazy`. Both modules
+  re-export `System.FS.API`.
 
 ### Non-breaking